pyproject.toml
[project]
name = "pokemon-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.12.6",
    "certifi>=2025.4.26",
    "google-genai>=1.18.0",
    "mcp[cli]>=1.9.2",
]
